Move bulk of atest main back into the direwolf package
Keeping cgo out of the cmds avoids namespace issues

Also move ACHAN2ADEV to cgo shims and dedupe
