+++ src/dtmf_test_shim.go
@@ -20,25 +20,20 @@ func push_button(channel int, button rune, ms int) {
 	C.push_button_test(C.int(channel), C.char(button), C.int(ms))
 }
 
-// #define ACHAN2ADEV(n) ((n)>>1)
-func ACHAN2ADEV(n int) int {
-	return n >> 1
-}
-
 func dtmf_test_main(t *testing.T) {
 	t.Helper()
 
 	var c = 0 // radio channel.
 	var my_audio_config C.struct_audio_s
 
-	my_audio_config.adev[ACHAN2ADEV(c)].defined = 1
-	my_audio_config.adev[ACHAN2ADEV(c)].samples_per_sec = 44100
+	my_audio_config.adev[ACHAN2ADEV(C.int(c))].defined = 1
+	my_audio_config.adev[ACHAN2ADEV(C.int(c))].samples_per_sec = 44100
 	my_audio_config.chan_medium[c] = C.MEDIUM_RADIO
 	my_audio_config.achan[c].dtmf_decode = C.DTMF_DECODE_ON
 
 	// Let's try to set up audio?
-	my_audio_config.adev[ACHAN2ADEV(c)].num_channels = 1
-	my_audio_config.adev[ACHAN2ADEV(c)].bits_per_sample = 8
+	my_audio_config.adev[ACHAN2ADEV(C.int(c))].num_channels = 1
+	my_audio_config.adev[ACHAN2ADEV(C.int(c))].bits_per_sample = 8
 	C.gen_tone_init(&my_audio_config, 100, 0)
 	C.ptt_init(&my_audio_config)
